requirements.txt
pytesseract
Pillow
opencv-python
fpdf